+++ tests/index.ts
@@ -1,9 +1,3 @@
-import { Identifiable } from '../src/SnapTypes';
-export interface TestData extends Identifiable {
-    [key: string]: any;
-    name: string;
-}
-
 export type DropEffect = 'none' | 'copy' | 'link' | 'move';
 export type EffectAllowed =
     | 'none'